Refactor F1 race command structure and add new functionality
- Renamed `f1_race` module to `f1` for better clarity and organization.
- Introduced new API and command handling for F1 race information, including fetching the F1 calendar and checking for upcoming races.
- Added embed creation for race announcements and improved command registration.
- Enhanced error handling and logging for race-related commands.
