+++ src/commands/mod.rs
@@ -1,4 +1,4 @@
 pub mod quote;
 pub mod scraper;
 pub mod version;
-pub mod f1_race;
+pub mod f1;
